requirements.txt
# Core dependencies
python-telegram-bot==20.7
openai==1.3.7
qdrant-client==1.7.0
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Environment and config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP requests
httpx==0.25.2
aiohttp==3.9.1

# Date/time handling
python-dateutil==2.8.2

# Markdown processing
markdown==3.5.1

# Logging
structlog==23.2.0

# Development
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-asyncio==0.21.1 